chore(shot): update to v4
